+++ EU4toV2/Source/Mappers/MinorityPops/MinorityPopMapper.cpp
@@ -26,25 +26,26 @@ void mappers::MinorityPopMapper::registerKeys()
 	registerRegex("[a-zA-Z0-9\\_.:]+", commonItems::ignoreItem);
 }
 
-bool mappers::MinorityPopMapper::matchMinorityPop(std::shared_ptr<V2::Pop> pop) const
+bool mappers::MinorityPopMapper::blankMajorityFromMinority(V2::Pop& pop) const
 {
+	// We are blanking majority part from a minority pop so that we may apply any
+	// religion to gipsies or any culture to mormons.
 	for (auto minorityItr : minorityPopMap)
 	{
-		if (pop->getCulture() == minorityItr.first && pop->getReligion() == minorityItr.second)
+		if (pop.getCulture() == minorityItr.first && pop.getReligion() == minorityItr.second)
 		{
 			return true;
 		}
-		if (minorityItr.first.empty() && pop->getReligion() == minorityItr.second)
+		if (minorityItr.first.empty() && pop.getReligion() == minorityItr.second)
 		{
-			pop->setCulture("");
+			pop.setCulture("");
 			return true;
 		}
-		if (pop->getCulture() == minorityItr.first && minorityItr.second.empty())
+		if (pop.getCulture() == minorityItr.first && minorityItr.second.empty())
 		{
-			pop->setReligion("");
+			pop.setReligion("");
 			return true;
 		}
 	}
-
 	return false;
 }
